+++ src/main/java/com/vacaro/benchlab/domain/ImplementationLanguage.java
@@ -7,4 +7,6 @@ public enum ImplementationLanguage {
     PYTHON,
     JAVA,
     GO,
+    RUST,
+    ASSEMBLY,
 }